+++ amazonorders/orders.py
@@ -77,15 +77,16 @@ class AmazonOrders:
 
                 if full_details:
                     if not order.order_details_link:
-                        logger.warning(f"order_details_link for Order {order.order_number} did not populate, "
-                                       f"cannot read full details.")
-
-                        continue
-
-                    self.amazon_session.get(order.order_details_link)
-                    order_details_tag = util.select_one(self.amazon_session.last_response_parsed,
-                                                        self.config.selectors.ORDER_DETAILS_ENTITY_SELECTOR)
-                    order = self.config.order_cls(order_details_tag, self.config, full_details=True, clone=order)
+                        logger.warning(f"Order {order.order_number} was partially populated, "
+                                       f"since order_details_link was not found.")
+                    elif len(util.select(order.parsed, self.config.selectors.ORDER_SKIP_ITEMS)) > 0:
+                        logger.warning(f"Order {order.order_number} was partially populated, "
+                                       f"since it is an unsupported Order type.")
+                    else:
+                        self.amazon_session.get(order.order_details_link)
+                        order_details_tag = util.select_one(self.amazon_session.last_response_parsed,
+                                                            self.config.selectors.ORDER_DETAILS_ENTITY_SELECTOR)
+                        order = self.config.order_cls(order_details_tag, self.config, full_details=True, clone=order)
 
                 orders.append(order)